Image_search/requirements.txt
pip==24.0
clip==0.2.0
chromadb==0.4.22
Flask==3.0.2
gunicorn==21.2.0
datasets==2.17.1
numpy==1.26.4
pandas==2.2.1
Pillow==10.0.1
requests==2.31.0
scikit-learn==1.3.0
transformers==4.37.2
torch==2.1.2
Werkzeug==3.0.1